fix(bmz-chart,bmz-gameplay): preserve bmson mine damage and sound
BMSON mine の音源 ID を damage として扱い、実際の浮動小数 damage と mine channel 音を失っていた。

- BMSON mine を専用の中間オブジェクトとして生成する。
- damage を浮動小数のまま gauge と描画snapshotまで運ぶ。
- mine channel 音をヒット時のキー音として再生する。
- 譜面指定音がない場合だけ既定 landmine SE へフォールバックする。

+++ crates/bmz-chart/src/import/bms_rs_adapter/objects.rs
@@ -107,7 +107,7 @@ pub(super) fn push_note_objects<T: KeyLayoutMapper>(
                 explicit_end_sound: false,
             },
             BmsNoteKind::Landmine => {
-                IntermediateObjectKind::MineNote { lane, wav_key: None, damage: wav_id }
+                IntermediateObjectKind::MineNote { lane, wav_key: None, damage: f64::from(wav_id) }
             }
         };
         objects.push(IntermediateObject {
