+++ crates/bmz-chart/src/model.rs
@@ -148,9 +148,9 @@ pub struct NoteEvent {
     pub tick: ChartTick,
     pub time: TimeUs,
     pub sound: Option<SoundId>,
-    /// Mine 専用のダメージ値（チャネル D系列に置かれた base36 値そのもの）。
+    /// Mine 専用のダメージ値。BMS は D/E 系列の整数値、BMSON は浮動小数。
     /// Mine 以外は常に None。
-    pub damage: Option<u16>,
+    pub damage: Option<f64>,
 }
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
